fix compilation warnings JIRA GP-36018
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102906 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/align/util/unit_test/genomic_compart_unit_test.cpp
@@ -46,7 +46,7 @@ USING_NCBI_SCOPE;
 USING_SCOPE(objects);
 
 struct SCompartmentData {
-    unsigned num_alignments;
+    size_t num_alignments;
     TSeqRange query_range;
     TSeqRange subject_range;
